Extract data source classes to separate files
